refactor: update _hooks to _utils
